Updated to Etyllica 3.3

+++ Big Circuit/src/br/com/longcircuit/gui/menu/MainMenu.java
@@ -23,7 +23,7 @@ import br.com.luvia.core.ApplicationGL;
 
 public class MainMenu extends ApplicationGL{
 
-	public MainMenu(float w, float h) {
+	public MainMenu(int w, int h) {
 		super(w,h);
 	}
